+++ vcell-core/src/main/java/cbit/vcell/export/server/datacontainer/ResultDataContainer.java
@@ -12,7 +12,7 @@ public interface ResultDataContainer {
 
     byte[] getDataBytes();
 
-    int getDataSize();
+    long getDataSize();
 
     ResultDataContainerID getId();
 }
